WIP: fix trip editor sidebar search review findings (checkpoint)
