+++ playwright.config.ts
@@ -6,6 +6,8 @@ export default defineConfig({
   testDir: './tests/e2e/trip-editor',
   outputDir: '.local/playwright/test-output',
   fullyParallel: false,
+  // Keep shared runbook trip checks serialized across split spec files.
+  workers: 1,
   forbidOnly: Boolean(process.env.CI),
   reporter: [['list'], ['html', { open: 'never', outputFolder: 'playwright-report' }]],
   use: {
